Moved the enum to standalone in the model package.
