+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/adapter/GBDeviceAdapter.java
@@ -11,9 +11,9 @@ import android.widget.TextView;
 
 import java.util.List;
 
-import nodomain.freeyourgadget.gadgetbridge.deviceevents.GBDeviceEventBatteryInfo;
 import nodomain.freeyourgadget.gadgetbridge.impl.GBDevice;
 import nodomain.freeyourgadget.gadgetbridge.R;
+import nodomain.freeyourgadget.gadgetbridge.model.BatteryState;
 
 public class GBDeviceAdapter extends ArrayAdapter<GBDevice> {
 
@@ -60,9 +60,9 @@ public class GBDeviceAdapter extends ArrayAdapter<GBDevice> {
         short batteryLevel = device.getBatteryLevel();
         if (batteryLevel != GBDevice.BATTERY_UNKNOWN) {
             batteryStatusLabel.setText("BAT: " + device.getBatteryLevel() + "%");
-            GBDeviceEventBatteryInfo.BatteryState batteryState = device.getBatteryState();
-            if (GBDeviceEventBatteryInfo.BatteryState.BATTERY_CHARGING.equals(batteryState) ||
-                    GBDeviceEventBatteryInfo.BatteryState.BATTERY_CHARGING_FULL.equals(batteryState)) {
+            BatteryState batteryState = device.getBatteryState();
+            if (BatteryState.BATTERY_CHARGING.equals(batteryState) ||
+                    BatteryState.BATTERY_CHARGING_FULL.equals(batteryState)) {
                 batteryStatusLabel.append(" CHG");
             }
         } else {
