+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/model/BatteryState.java
@@ -1,0 +1,10 @@
+package nodomain.freeyourgadget.gadgetbridge.model;
+
+public enum BatteryState {
+    UNKNOWN,
+    BATTERY_NORMAL,
+    BATTERY_LOW,
+    BATTERY_CHARGING,
+    BATTERY_CHARGING_FULL,
+    BATTERY_NOT_CHARGING_FULL
+}
